Split home route data helpers
