+++ tests/home-route-architecture.test.ts
@@ -6,7 +6,19 @@ import test from 'node:test';
 const root = process.cwd();
 const pagePath = join(root, 'frontend/app/(localized)/[locale]/(marketing)/(home)/page.tsx');
 const routeDataPath = join(root, 'frontend/app/(localized)/[locale]/(marketing)/(home)/_lib/home-route-data.ts');
+const routeDataDir = join(root, 'frontend/app/(localized)/[locale]/(marketing)/(home)/_lib/home-route-data');
 const jsonLdPath = join(root, 'frontend/app/(localized)/[locale]/(marketing)/(home)/_lib/home-jsonld.ts');
+const routeDataModules = [
+  'best-for.ts',
+  'comparisons.ts',
+  'constants.ts',
+  'engine-stats.ts',
+  'examples.ts',
+  'filters.ts',
+  'formatting.ts',
+  'hero.ts',
+  'types.ts',
+];
 
 const pageSource = readFileSync(pagePath, 'utf8');
 const routeDataSource = readFileSync(routeDataPath, 'utf8');
@@ -37,7 +49,7 @@ test('home route keeps page.tsx as a thin orchestrator', () => {
     'const HOMEPAGE_EXAMPLE_VIDEO_OVERRIDES',
   ]) {
     assert.doesNotMatch(pageSource, new RegExp(owner), `${owner} should not be owned by home/page.tsx`);
-    assert.match(routeDataSource, new RegExp(owner), `${owner} should live in the route-local data module`);
+    assert.doesNotMatch(routeDataSource, new RegExp(owner), `${owner} should not be owned by the home-route-data facade`);
   }
 
   for (const owner of ['function buildSoftwareSchema', 'function buildFaqSchema', 'function serializeJsonLd']) {
@@ -63,10 +75,39 @@ test('home route data module exposes the orchestration contract explicitly', () 
   ]) {
     assert.match(
       routeDataSource,
-      new RegExp(`export (?:async )?(?:function|const) ${exportedName}\\b`),
+      new RegExp(`export \\{[\\s\\S]*${exportedName}`),
       `${exportedName} should be exported for the home route orchestrator`
     );
   }
+
+  assert.match(routeDataSource, /export type \{ EngineStats, RedesignContent \}/);
+
+  const facadeLineCount = routeDataSource.split('\n').length;
+  assert.ok(facadeLineCount <= 40, `home-route-data.ts should stay a thin facade, got ${facadeLineCount}`);
+});
+
+test('home route data responsibilities are split into focused modules', () => {
+  for (const moduleName of routeDataModules) {
+    const modulePath = join(routeDataDir, moduleName);
+    assert.ok(existsSync(modulePath), `${moduleName} should exist under home-route-data`);
+    const moduleSource = readFileSync(modulePath, 'utf8');
+    const lineCount = moduleSource.split('\n').length;
+    assert.ok(lineCount <= 240, `${moduleName} should stay below 240 lines, got ${lineCount}`);
+  }
+
+  const typesSource = readFileSync(join(routeDataDir, 'types.ts'), 'utf8');
+  const engineStatsSource = readFileSync(join(routeDataDir, 'engine-stats.ts'), 'utf8');
+  const heroSource = readFileSync(join(routeDataDir, 'hero.ts'), 'utf8');
+  const examplesSource = readFileSync(join(routeDataDir, 'examples.ts'), 'utf8');
+  const comparisonsSource = readFileSync(join(routeDataDir, 'comparisons.ts'), 'utf8');
+  const constantsSource = readFileSync(join(routeDataDir, 'constants.ts'), 'utf8');
+
+  assert.match(typesSource, /export type RedesignContent =/);
+  assert.match(engineStatsSource, /export function computeEngineStats/);
+  assert.match(examplesSource, /export async function loadHomepageExamples/);
+  assert.match(comparisonsSource, /export async function buildComparisonCardsWithExampleMedia/);
+  assert.match(heroSource, /export function buildProgrammedHeroItems/);
+  assert.match(constantsSource, /export const HOMEPAGE_EXAMPLE_VIDEO_OVERRIDES/);
 });
 
 test('home JSON-LD module owns schema serialization helpers', () => {
